requirements.txt
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1610094676799/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
certifi==2021.10.8
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1627074992779/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
greenlet==1.1.2
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1634726430246/work/dist/ipykernel-6.4.2-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1632764055494/work
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1610146833874/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1631852743958/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
numpy==1.21.3
pandas==1.3.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
prettytable==0.7.2
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1634921890410/work
psycopg2-binary==2.9.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1631793589817/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy==1.4.26
sqlparse==0.4.2
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1625489059091/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1630423529112/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
